Split expressions.rs into calls.rs, collections.rs, control.rs
