+++ src/codegen/emit_wasm/expressions.rs
@@ -1,14 +1,22 @@
 //! IrExpr → WASM instruction emission.
 
-use crate::ir::{BinOp, CallTarget, IrExpr, IrExprKind, IrMatchArm, IrPattern, UnOp};
+use crate::ir::{BinOp, IrExpr, IrExprKind, UnOp};
 use crate::types::Ty;
 use wasm_encoder::{BlockType, Instruction, MemArg, ValType};
 
 use super::FuncCompiler;
 use super::values;
 
-fn mem(offset: u64) -> MemArg {
+pub(super) fn mem(offset: u64) -> MemArg {
     MemArg { offset, align: 2, memory_index: 0 }
+}
+
+#[derive(Clone, Copy)]
+pub(super) enum CmpKind {
+    Lt,
+    Gt,
+    Lte,
+    Gte,
 }
 
 impl FuncCompiler<'_> {
@@ -353,7 +361,7 @@ impl FuncCompiler<'_> {
         }
     }
 
-    fn emit_binop(&mut self, op: BinOp, left: &IrExpr, right: &IrExpr) {
+    pub(super) fn emit_binop(&mut self, op: BinOp, left: &IrExpr, right: &IrExpr) {
         match op {
             // ── Arithmetic ──
             BinOp::AddInt => {
@@ -481,7 +489,7 @@ impl FuncCompiler<'_> {
         }
     }
 
-    fn emit_eq(&mut self, left: &IrExpr, right: &IrExpr, negate: bool) {
+    pub(super) fn emit_eq(&mut self, left: &IrExpr, right: &IrExpr, negate: bool) {
         self.emit_expr(left);
         self.emit_expr(right);
         match &left.ty {
@@ -539,7 +547,7 @@ impl FuncCompiler<'_> {
         }
     }
 
-    fn emit_cmp_instruction(&mut self, ty: &Ty, kind: CmpKind) {
+    pub(super) fn emit_cmp_instruction(&mut self, ty: &Ty, kind: CmpKind) {
         match (ty, kind) {
             (Ty::Int, CmpKind::Lt) => { self.func.instruction(&Instruction::I64LtS); }
             (Ty::Int, CmpKind::Gt) => { self.func.instruction(&Instruction::I64GtS); }
@@ -553,745 +561,9 @@ impl FuncCompiler<'_> {
         }
     }
 
-    fn emit_call(&mut self, target: &CallTarget, args: &[IrExpr], _ret_ty: &Ty) {
-        match target {
-            CallTarget::Named { name } => {
-                match name.as_str() {
-                    "println" => {
-                        let arg = &args[0];
-                        match &arg.ty {
-                            Ty::String => {
-                                self.emit_expr(arg);
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.println_str));
-                            }
-                            Ty::Int => {
-                                self.emit_expr(arg);
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.println_int));
-                            }
-                            Ty::Bool => {
-                                // Convert bool to "true"/"false"
-                                self.emit_expr(arg);
-                                let true_str = self.emitter.intern_string("true");
-                                let false_str = self.emitter.intern_string("false");
-                                self.func.instruction(&Instruction::If(BlockType::Result(wasm_encoder::ValType::I32)));
-                                self.func.instruction(&Instruction::I32Const(true_str as i32));
-                                self.func.instruction(&Instruction::Else);
-                                self.func.instruction(&Instruction::I32Const(false_str as i32));
-                                self.func.instruction(&Instruction::End);
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.println_str));
-                            }
-                            Ty::Float => {
-                                // Phase 1: print float as int (truncated)
-                                self.emit_expr(arg);
-                                self.func.instruction(&Instruction::I64TruncF64S);
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.println_int));
-                            }
-                            _ => {
-                                // Unsupported type: skip arg and print "<unsupported>"
-                                let s = self.emitter.intern_string("<unsupported>");
-                                self.func.instruction(&Instruction::I32Const(s as i32));
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.println_str));
-                            }
-                        }
-                    }
-                    "assert_eq" => {
-                        self.emit_assert_eq(&args[0], &args[1]);
-                    }
-                    "assert" => {
-                        // assert(cond) or assert(cond, msg) — trap if false
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::I32Eqz);
-                        self.func.instruction(&Instruction::If(BlockType::Empty));
-                        self.func.instruction(&Instruction::Unreachable);
-                        self.func.instruction(&Instruction::End);
-                        // Drop message arg if present (evaluated but unused)
-                    }
-                    "assert_ne" => {
-                        // assert_ne(left, right) — trap if equal
-                        self.emit_eq(&args[0], &args[1], false);
-                        // If equal → trap
-                        self.func.instruction(&Instruction::If(BlockType::Empty));
-                        self.func.instruction(&Instruction::Unreachable);
-                        self.func.instruction(&Instruction::End);
-                    }
-                    _ => {
-                        // Check if this is a unit variant constructor (e.g., Red, None)
-                        if args.is_empty() {
-                            if let Some(tag) = self.find_unit_variant_tag(name) {
-                                // Allocate [tag:i32]
-                                self.func.instruction(&Instruction::I32Const(4));
-                                self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-                                let scratch = self.match_i32_base + self.match_depth;
-                                self.func.instruction(&Instruction::LocalSet(scratch));
-                                self.func.instruction(&Instruction::LocalGet(scratch));
-                                self.func.instruction(&Instruction::I32Const(tag as i32));
-                                self.func.instruction(&Instruction::I32Store(MemArg {
-                                    offset: 0, align: 2, memory_index: 0,
-                                }));
-                                self.func.instruction(&Instruction::LocalGet(scratch));
-                                return;
-                            }
-                        }
-                        // User-defined function call
-                        for arg in args {
-                            self.emit_expr(arg);
-                        }
-                        if let Some(&func_idx) = self.emitter.func_map.get(name.as_str()) {
-                            self.func.instruction(&Instruction::Call(func_idx));
-                        } else {
-                            self.func.instruction(&Instruction::Unreachable);
-                        }
-                    }
-                }
-            }
-
-            CallTarget::Module { module, func } => {
-                match (module.as_str(), func.as_str()) {
-                    ("int", "to_string") => {
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    ("float", "to_string") => {
-                        // Phase 1: truncate to int, then int_to_string
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::I64TruncF64S);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    ("string", "length") | ("string", "len") => {
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::I32Load(mem(0)));
-                        self.func.instruction(&Instruction::I64ExtendI32U);
-                    }
-                    ("int", "parse") => {
-                        // Stub: return ok(0) as Result[Int, String]
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::Drop);
-                        // Allocate Result: [tag=0 (ok), value=0 (i64)]
-                        self.func.instruction(&Instruction::I32Const(12)); // 4 tag + 8 i64
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-                        let scratch = self.match_i32_base + self.match_depth;
-                        self.func.instruction(&Instruction::LocalSet(scratch));
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.func.instruction(&Instruction::I32Const(0)); // tag = ok
-                        self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.func.instruction(&Instruction::I64Const(0)); // value = 0
-                        self.func.instruction(&Instruction::I64Store(MemArg { offset: 4, align: 3, memory_index: 0 }));
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                    }
-                    ("string", "contains") => {
-                        // string.contains(haystack, needle) -> bool
-                        // Brute force: O(n*m) substring search
-                        self.emit_expr(&args[0]); // haystack ptr
-                        self.emit_expr(&args[1]); // needle ptr
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.str_contains));
-                    }
-                    ("string", "trim") => {
-                        // Stub: return the string as-is (no whitespace handling)
-                        self.emit_expr(&args[0]);
-                    }
-                    ("string", "to_upper") | ("string", "to_lower") => {
-                        // Stub: return as-is
-                        self.emit_expr(&args[0]);
-                    }
-                    ("string", "repeat") | ("string", "reverse") | ("string", "replace")
-                    | ("string", "split") | ("string", "join") | ("string", "slice")
-                    | ("string", "get") | ("string", "count") | ("string", "starts_with")
-                    | ("string", "ends_with") | ("string", "index_of")
-                    | ("string", "pad_start") | ("string", "pad_end")
-                    | ("string", "trim_start") | ("string", "trim_end") => {
-                        self.emit_stub_call(args);
-                    }
-                    ("list", "map") => {
-                        self.emit_list_map(&args[0], &args[1], _ret_ty);
-                    }
-                    ("list", "filter") => {
-                        self.emit_list_filter(&args[0], &args[1], _ret_ty);
-                    }
-                    ("list", "fold") => {
-                        // fold(list, init, fn(acc, elem) -> acc)
-                        self.emit_list_fold(&args[0], &args[1], &args[2]);
-                    }
-                    ("list", "reverse") => {
-                        self.emit_list_reverse(&args[0], _ret_ty);
-                    }
-                    ("list", "find") | ("list", "any") | ("list", "all")
-                    | ("list", "count") | ("list", "sort_by") | ("list", "flat_map")
-                    | ("list", "filter_map") | ("list", "get") | ("list", "drop")
-                    | ("list", "take") | ("list", "zip")
-                    | ("list", "enumerate") | ("list", "contains") | ("list", "sort") => {
-                        self.emit_stub_call(args);
-                    }
-                    ("map", "len") | ("map", "length") | ("map", "size") => {
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::I32Load(mem(0)));
-                        self.func.instruction(&Instruction::I64ExtendI32U);
-                    }
-                    ("list", "len") | ("list", "length") => {
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::I32Load(mem(0)));
-                        self.func.instruction(&Instruction::I64ExtendI32U);
-                    }
-                    _ => {
-                        self.emit_stub_call(args);
-                    }
-                }
-            }
-
-            CallTarget::Method { object, method } => {
-                // UFCS method calls: obj.method(args)
-                match method.as_str() {
-                    "to_string" if matches!(object.ty, Ty::Int) => {
-                        self.emit_expr(object);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    "len" | "length" | "string.len" | "list.len" | "map.len" => {
-                        // .len() for String, List, Map — all store length at offset 0
-                        self.emit_expr(object);
-                        self.func.instruction(&Instruction::I32Load(mem(0)));
-                        self.func.instruction(&Instruction::I64ExtendI32U);
-                    }
-                    "to_string" if matches!(object.ty, Ty::Float) => {
-                        self.emit_expr(object);
-                        self.func.instruction(&Instruction::I64TruncF64S);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    "contains" if matches!(object.ty, Ty::String) => {
-                        self.emit_expr(object);
-                        self.emit_expr(&args[0]);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.str_contains));
-                    }
-                    _ => {
-                        self.emit_expr(object);
-                        if values::ty_to_valtype(&object.ty).is_some() {
-                            self.func.instruction(&Instruction::Drop);
-                        }
-                        self.emit_stub_call(args);
-                    }
-                }
-            }
-
-            CallTarget::Computed { callee } => {
-                // Closure call: callee is a closure ptr [table_idx: i32][env_ptr: i32]
-                // Stack order for call_indirect: [env_ptr, args..., table_idx]
-                let scratch = self.match_i32_base + self.match_depth;
-
-                // Evaluate callee → closure ptr
-                self.emit_expr(callee);
-                self.func.instruction(&Instruction::LocalSet(scratch));
-
-                // Push env_ptr (first hidden arg)
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Load(MemArg {
-                    offset: 4, align: 2, memory_index: 0,
-                }));
-
-                // Push declared args
-                for arg in args {
-                    self.emit_expr(arg);
-                }
-
-                // Push table_idx (on top of stack for call_indirect)
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Load(MemArg {
-                    offset: 0, align: 2, memory_index: 0,
-                }));
-
-                // Closure calling convention type: (env: i32, params...) -> ret
-                if let Ty::Fn { params, ret } = &callee.ty {
-                    let mut closure_params = vec![ValType::I32]; // env_ptr
-                    for p in params {
-                        if let Some(vt) = values::ty_to_valtype(p) {
-                            closure_params.push(vt);
-                        }
-                    }
-                    let ret_types = values::ret_type(ret);
-                    let type_idx = self.emitter.register_type(closure_params, ret_types);
-                    self.func.instruction(&Instruction::CallIndirect {
-                        type_index: type_idx,
-                        table_index: 0,
-                    });
-                } else {
-                    self.func.instruction(&Instruction::Unreachable);
-                }
-            }
-        }
-    }
-
-    /// Emit a FnRef as a closure: allocate [wrapper_table_idx, 0] on heap.
-    fn emit_fn_ref_closure(&mut self, name: &str) {
-        if let Some(&wrapper_table_idx) = self.emitter.fn_ref_wrappers.get(name) {
-            // Allocate closure: [table_idx: i32][env_ptr: i32] = 8 bytes
-            self.func.instruction(&Instruction::I32Const(8));
-            self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-            let scratch = self.match_i32_base + self.match_depth;
-            self.func.instruction(&Instruction::LocalSet(scratch));
-
-            // Store table_idx
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.func.instruction(&Instruction::I32Const(wrapper_table_idx as i32));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-            // Store env_ptr = 0
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.func.instruction(&Instruction::I32Const(0));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 4, align: 2, memory_index: 0 }));
-
-            // Return closure ptr
-            self.func.instruction(&Instruction::LocalGet(scratch));
-        } else {
-            eprintln!("WARNING: FnRef wrapper not found for '{}', using direct table entry", name);
-            self.func.instruction(&Instruction::Unreachable);
-        }
-    }
-
-    /// Emit a lambda as a closure: allocate env + closure on heap.
-    fn emit_lambda_closure(&mut self, _params: &[(crate::ir::VarId, Ty)], _body: &IrExpr) {
-        let lambda_idx = self.emitter.lambda_counter.get();
-        self.emitter.lambda_counter.set(lambda_idx + 1);
-
-        if lambda_idx >= self.emitter.lambdas.len() {
-            self.func.instruction(&Instruction::Unreachable);
-            return;
-        }
-
-        let table_idx = self.emitter.lambdas[lambda_idx].table_idx;
-        let captures = self.emitter.lambdas[lambda_idx].captures.clone();
-
-        let scratch = self.match_i32_base + self.match_depth;
-
-        if captures.is_empty() {
-            // No captures: allocate closure [table_idx, 0]
-            self.func.instruction(&Instruction::I32Const(8));
-            self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-            self.func.instruction(&Instruction::LocalSet(scratch));
-
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.func.instruction(&Instruction::I32Const(table_idx as i32));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.func.instruction(&Instruction::I32Const(0));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 4, align: 2, memory_index: 0 }));
-
-            self.func.instruction(&Instruction::LocalGet(scratch));
-        } else {
-            // Allocate env: each capture gets 8 bytes (padded for alignment)
-            let env_size = (captures.len() as u32) * 8;
-            self.func.instruction(&Instruction::I32Const(env_size as i32));
-            self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-            let env_scratch = scratch; // reuse for env_ptr
-            self.func.instruction(&Instruction::LocalSet(env_scratch));
-
-            // Store each captured variable into env
-            for (ci, (vid, ty)) in captures.iter().enumerate() {
-                let offset = (ci as u32) * 8;
-                self.func.instruction(&Instruction::LocalGet(env_scratch));
-                if let Some(&local_idx) = self.var_map.get(&vid.0) {
-                    self.func.instruction(&Instruction::LocalGet(local_idx));
-                } else {
-                    // Variable not in scope — shouldn't happen
-                    self.func.instruction(&Instruction::I32Const(0));
-                }
-                self.emit_store_at(ty, offset);
-            }
-
-            // Allocate closure: [table_idx, env_ptr]
-            self.func.instruction(&Instruction::I32Const(8));
-            self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-            let closure_scratch = scratch + 1; // second i32 scratch slot
-            self.func.instruction(&Instruction::LocalSet(closure_scratch));
-
-            self.func.instruction(&Instruction::LocalGet(closure_scratch));
-            self.func.instruction(&Instruction::I32Const(table_idx as i32));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-            self.func.instruction(&Instruction::LocalGet(closure_scratch));
-            self.func.instruction(&Instruction::LocalGet(env_scratch));
-            self.func.instruction(&Instruction::I32Store(MemArg { offset: 4, align: 2, memory_index: 0 }));
-
-            self.func.instruction(&Instruction::LocalGet(closure_scratch));
-        }
-    }
-
-    /// Emit list.map(list, fn) → new list.
-    /// Uses memory scratch [8..16] as temp storage for the mapped value.
-    fn emit_list_map(&mut self, list_arg: &IrExpr, fn_arg: &IrExpr, ret_ty: &Ty) {
-        // Determine element types
-        let in_elem_ty = if let Ty::Applied(_, args) = &list_arg.ty {
-            args.first().cloned().unwrap_or(Ty::Int)
-        } else { Ty::Int };
-        let out_elem_ty = if let Ty::Applied(_, args) = ret_ty {
-            args.first().cloned().unwrap_or(Ty::Int)
-        } else { Ty::Int };
-        let in_size = values::byte_size(&in_elem_ty);
-        let out_size = values::byte_size(&out_elem_ty);
-
-        let s = self.match_i32_base + self.match_depth;
-        // s+0=src, s+1=fn, s+2=dst, s+3=len, s+4=i
-        // Need 5 scratch i32 slots — use what we have + i64 scratch
-        let src = s;
-        let fn_ptr = s + 1;
-        let dst = self.match_i64_base + self.match_depth; // borrow i64 scratch as i32
-        // Actually this won't work with typed locals. Let me use a simpler approach:
-        // Store everything in i32 scratch slots. Need to ensure enough scratch.
-        // For now, just use s, s+1 for src and fn, and allocate dst inline.
-
-        // Eval src list
-        self.emit_expr(list_arg);
-        self.func.instruction(&Instruction::LocalSet(src));
-        // Eval fn closure
-        self.emit_expr(fn_arg);
-        self.func.instruction(&Instruction::LocalSet(fn_ptr));
-
-        // Get length
-        self.func.instruction(&Instruction::LocalGet(src));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        // Allocate output: [len + len * out_size]
-        // Store len in dst after alloc
-        // Stack: [len]
-        // Dup len: local.tee to a temp... use fn_ptr as temp? No, it holds fn.
-        // Just reload len later
-        self.func.instruction(&Instruction::Drop); // drop len for now
-
-        // Alloc output: 4 + src.len * out_size
-        self.func.instruction(&Instruction::I32Const(4));
-        self.func.instruction(&Instruction::LocalGet(src));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        self.func.instruction(&Instruction::I32Const(out_size as i32));
-        self.func.instruction(&Instruction::I32Mul);
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-        // dst ptr on stack — store it. We need another scratch.
-        // Use the i64 scratch slot, but store i32 via I32 store... won't work with typed locals.
-        // Solution: just use s+1 for dst after we're done with fn_ptr.
-        // Actually we still need fn_ptr in the loop. Let me rethink.
-
-        // Simpler approach: store src and fn to memory scratch area (offset 0-8),
-        // use locals only for loop counter and dst.
-        // Actually, let me just use more scratch. The scratch_depth should be enough.
-
-        // OK simplest: use 2 i32 scratch (src, fn). Store dst on stack, use tee.
-        // This is getting complex. Let me emit a runtime call instead.
-
-        // RUNTIME APPROACH: __list_map is hard because of closure calling convention.
-        // Let me inline it with simple structure.
-
-        // Reset approach: I'll allocate a result list, then loop calling the closure.
-        self.func.instruction(&Instruction::Drop); // drop alloc result, redo properly
-
-        // Re-eval src length
-        self.func.instruction(&Instruction::LocalGet(src));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        // Use i32.mul for total output size, add 4 for len prefix
-        let out_total_local = self.match_i64_base + self.match_depth; // reuse i64 as i32
-        // Can't reuse i64 local for i32 value. Need proper i32.
-        // Fallback: just rebuild using nested alloc + inline loop with break.
-
-        // === Clean implementation ===
-        // I need: src_ptr, fn_ptr, dst_ptr, len, idx. That's 5 i32 values.
-        // I have 2 i32 scratch slots (s, s+1). Not enough.
-        // But I can use memory scratch area (offset 0-15 is already scratch for fd_write).
-
-        // Store src_ptr at mem[0], fn_ptr at mem[4]
-        self.func.instruction(&Instruction::Drop); // clean stack
-
-        self.func.instruction(&Instruction::I32Const(0));
-        self.func.instruction(&Instruction::LocalGet(src));
-        self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        self.func.instruction(&Instruction::I32Const(4));
-        self.func.instruction(&Instruction::LocalGet(fn_ptr));
-        self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-        // len = src.len
-        self.func.instruction(&Instruction::LocalGet(src));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        self.func.instruction(&Instruction::LocalSet(src)); // reuse src as len
-
-        // Alloc dst: 4 + len * out_size
-        self.func.instruction(&Instruction::I32Const(4));
-        self.func.instruction(&Instruction::LocalGet(src)); // len
-        self.func.instruction(&Instruction::I32Const(out_size as i32));
-        self.func.instruction(&Instruction::I32Mul);
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-        self.func.instruction(&Instruction::LocalSet(fn_ptr)); // reuse fn_ptr as dst
-
-        // Store len in dst
-        self.func.instruction(&Instruction::LocalGet(fn_ptr));
-        self.func.instruction(&Instruction::LocalGet(src)); // len
-        self.func.instruction(&Instruction::I32Store(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-        // src_ptr = mem[0], fn_closure = mem[4], len = src local, dst = fn_ptr local
-        // idx = 0 (use s+1 scratch)
-        let idx = s + 1;
-        self.func.instruction(&Instruction::I32Const(0));
-        self.func.instruction(&Instruction::LocalSet(idx));
-
-        // Loop
-        self.func.instruction(&Instruction::Block(BlockType::Empty));
-        self.func.instruction(&Instruction::Loop(BlockType::Empty));
-        let saved_depth = self.depth;
-        self.depth += 2;
-
-        // Break if idx >= len
-        self.func.instruction(&Instruction::LocalGet(idx));
-        self.func.instruction(&Instruction::LocalGet(src)); // len
-        self.func.instruction(&Instruction::I32GeU);
-        self.func.instruction(&Instruction::BrIf(1));
-
-        // Load fn closure from mem[4]
-        self.func.instruction(&Instruction::I32Const(4));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        // This is closure ptr. Load env_ptr from closure[4]
-        let closure_scratch = s; // reuse s temporarily
-        self.func.instruction(&Instruction::LocalSet(closure_scratch));
-
-        // Call: env_ptr, element, table_idx → call_indirect
-        // env_ptr
-        self.func.instruction(&Instruction::LocalGet(closure_scratch));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 4, align: 2, memory_index: 0 }));
-
-        // Load element from src: mem[0] is src_ptr, element at src_ptr + 4 + idx * in_size
-        self.func.instruction(&Instruction::I32Const(0));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-        self.func.instruction(&Instruction::I32Const(4));
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::LocalGet(idx));
-        self.func.instruction(&Instruction::I32Const(in_size as i32));
-        self.func.instruction(&Instruction::I32Mul);
-        self.func.instruction(&Instruction::I32Add);
-        self.emit_load_at(&in_elem_ty, 0);
-
-        // table_idx
-        self.func.instruction(&Instruction::LocalGet(closure_scratch));
-        self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-
-        // call_indirect with closure calling convention
-        if let Ty::Fn { params, ret } = &fn_arg.ty {
-            let mut closure_params = vec![ValType::I32]; // env
-            for p in params {
-                if let Some(vt) = values::ty_to_valtype(p) { closure_params.push(vt); }
-            }
-            let ret_types = values::ret_type(ret);
-            let type_idx = self.emitter.register_type(closure_params, ret_types);
-            self.func.instruction(&Instruction::CallIndirect { type_index: type_idx, table_index: 0 });
-        } else {
-            self.func.instruction(&Instruction::Unreachable);
-        }
-
-        // Store result to memory scratch [8], then copy to dst
-        // Result is on stack (typed: out_elem_ty). Store to mem[8] temporarily.
-        self.func.instruction(&Instruction::I32Const(8));
-        // Stack: [result, addr=8]. But store needs [addr, value]. Swap using scratch.
-        // Actually, I32Const 8 is UNDER the result. WASM evaluates left to right.
-        // So stack is [..., result, 8]. That's wrong for store.
-        // Need: I32Const(8) first, then result. But result is already computed.
-        // Use memory scratch differently: store result via a temp.
-        // Simplest: use I32Const(8) as addr with offset in store instruction.
-        self.func.instruction(&Instruction::Drop); // drop the 8
-        // Store result to mem[8..16] using I32Store/I64Store at absolute addr 8
-        match values::ty_to_valtype(&out_elem_ty) {
-            Some(ValType::I64) => {
-                // result (i64) on stack. Store to mem[8].
-                // Need addr on stack first. But result is on top. Use scratch local.
-                self.func.instruction(&Instruction::LocalSet(src)); // temp store result in src (reused as i32... type mismatch!)
-                // Can't store i64 in i32 local. Use i64 scratch.
-                // This is impossible without a proper i64 temp local.
-                // HACK: store directly to memory using a known addr
-                // Actually: i32.const 8; <value already consumed>; i64.store
-                // The problem is value was already consumed by LocalSet.
-                // Let me restructure: compute addr first, then call.
-            }
-            _ => {}
-        }
-        // This approach doesn't work. Let me restructure the entire loop.
-        // Compute dst addr BEFORE the call, store addr in idx temp, call, then store.
-
-        // ... actually let me just compute addr, call, then use memory as bridge:
-        // Simplest correct approach: reverse the computation order.
-        // Before the call: compute and store dst addr to mem[8]
-        // Then call (result on stack)
-        // Then load dst addr from mem[8], swap not needed because I can use store with offset
-
-        // FORGET ALL THE ABOVE. Let me just rewrite the whole loop cleanly.
-
-        // idx++
-        self.func.instruction(&Instruction::LocalGet(idx));
-        self.func.instruction(&Instruction::I32Const(1));
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::LocalSet(idx));
-        self.func.instruction(&Instruction::Br(0));
-
-        self.depth = saved_depth;
-        self.func.instruction(&Instruction::End);
-        self.func.instruction(&Instruction::End);
-
-        // Return dst
-        self.func.instruction(&Instruction::LocalGet(fn_ptr));
-    }
-
-    /// Emit list.filter — stub for now
-    fn emit_list_filter(&mut self, _list_arg: &IrExpr, _fn_arg: &IrExpr, _ret_ty: &Ty) {
-        self.emit_stub_call(&[_list_arg.clone(), _fn_arg.clone()]);
-    }
-
-    /// Emit list.fold(list, init, fn(acc, elem) -> acc)
-    fn emit_list_fold(&mut self, _list_arg: &IrExpr, _init_arg: &IrExpr, _fn_arg: &IrExpr) {
-        self.emit_stub_call(&[_list_arg.clone(), _init_arg.clone(), _fn_arg.clone()]);
-    }
-
-    /// Emit list.reverse
-    fn emit_list_reverse(&mut self, _list_arg: &IrExpr, _ret_ty: &Ty) {
-        self.emit_stub_call(&[_list_arg.clone()]);
-    }
-
-    /// Emit a stub for an unimplemented call: evaluate args (for side effects), drop values, unreachable.
-    fn emit_stub_call(&mut self, args: &[IrExpr]) {
-        for arg in args {
-            self.emit_expr(arg);
-            // Only drop if the arg produces a value
-            if values::ty_to_valtype(&arg.ty).is_some() {
-                self.func.instruction(&Instruction::Drop);
-            }
-        }
-        self.func.instruction(&Instruction::Unreachable);
-    }
-
-    /// Emit assert_eq(left, right): compare values, trap if not equal.
-    fn emit_assert_eq(&mut self, left: &IrExpr, right: &IrExpr) {
-        // Use the same equality logic as BinOp::Eq
-        self.emit_eq(left, right, false);
-        // If not equal (result == 0), trap
-        self.func.instruction(&Instruction::I32Eqz);
-        self.func.instruction(&Instruction::If(BlockType::Empty));
-        self.func.instruction(&Instruction::Unreachable);
-        self.func.instruction(&Instruction::End);
-    }
-
-    /// Concatenate two strings on the heap via __concat_str runtime.
-    fn emit_concat_str(&mut self, left: &IrExpr, right: &IrExpr) {
-        self.emit_expr(left);
-        self.emit_expr(right);
-        self.func.instruction(&Instruction::Call(self.emitter.rt.concat_str));
-    }
-
-    /// String interpolation: convert each part to string, then concat.
-    fn emit_string_interp(&mut self, parts: &[crate::ir::IrStringPart]) {
-        if parts.is_empty() {
-            let empty = self.emitter.intern_string("");
-            self.func.instruction(&Instruction::I32Const(empty as i32));
-            return;
-        }
-
-        // Emit first part as a string
-        self.emit_string_part(&parts[0]);
-
-        // For each subsequent part: emit it, then concat with accumulator
-        for part in &parts[1..] {
-            self.emit_string_part(part);
-            self.func.instruction(&Instruction::Call(self.emitter.rt.concat_str));
-        }
-    }
-
-    /// Emit a single string interpolation part as a string (i32 pointer).
-    fn emit_string_part(&mut self, part: &crate::ir::IrStringPart) {
-        match part {
-            crate::ir::IrStringPart::Lit { value } => {
-                let offset = self.emitter.intern_string(value);
-                self.func.instruction(&Instruction::I32Const(offset as i32));
-            }
-            crate::ir::IrStringPart::Expr { expr } => {
-                match &expr.ty {
-                    Ty::String => self.emit_expr(expr),
-                    Ty::Int => {
-                        self.emit_expr(expr);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    Ty::Bool => {
-                        self.emit_expr(expr);
-                        let t = self.emitter.intern_string("true");
-                        let f = self.emitter.intern_string("false");
-                        self.func.instruction(&Instruction::If(wasm_encoder::BlockType::Result(wasm_encoder::ValType::I32)));
-                        self.func.instruction(&Instruction::I32Const(t as i32));
-                        self.func.instruction(&Instruction::Else);
-                        self.func.instruction(&Instruction::I32Const(f as i32));
-                        self.func.instruction(&Instruction::End);
-                    }
-                    Ty::Float => {
-                        self.emit_expr(expr);
-                        self.func.instruction(&Instruction::I64TruncF64S);
-                        self.func.instruction(&Instruction::Call(self.emitter.rt.int_to_string));
-                    }
-                    _ => {
-                        // Fallback: emit the expression (already a string pointer or unsupported)
-                        self.emit_expr(expr);
-                    }
-                }
-            }
-        }
-    }
-    /// Emit a record/variant construction: allocate memory, store fields.
-    /// For variants (detected via name + type), prepends a tag i32 before fields.
-    fn emit_record(&mut self, name: Option<&str>, fields: &[(String, IrExpr)], result_ty: &Ty) {
-        // Check if this is a variant constructor
-        let tag = self.resolve_variant_tag(name, result_ty);
-        let tag_size: u32 = if tag.is_some() { 4 } else { 0 };
-
-        // Compute field types and total size
-        let field_types: Vec<(String, Ty)> = fields.iter()
-            .map(|(n, expr)| (n.clone(), expr.ty.clone()))
-            .collect();
-        let total_size = tag_size + values::record_size(&field_types);
-
-        // Allocate
-        self.func.instruction(&Instruction::I32Const(total_size as i32));
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-
-        let scratch = self.match_i32_base + self.match_depth;
-        self.func.instruction(&Instruction::LocalSet(scratch));
-
-        // Write tag if variant
-        if let Some(tag_val) = tag {
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.func.instruction(&Instruction::I32Const(tag_val as i32));
-            self.func.instruction(&Instruction::I32Store(MemArg {
-                offset: 0, align: 2, memory_index: 0,
-            }));
-        }
-
-        // Store each field (offset starts after tag)
-        let mut offset = tag_size;
-        for (_, field_expr) in fields {
-            let field_size = values::byte_size(&field_expr.ty);
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.emit_expr(field_expr);
-            self.emit_store_at(&field_expr.ty, offset);
-            offset += field_size;
-        }
-
-        // Return ptr
-        self.func.instruction(&Instruction::LocalGet(scratch));
-    }
-
-    /// Look up variant tag for a constructor name within a variant type.
-    fn resolve_variant_tag(&self, name: Option<&str>, result_ty: &Ty) -> Option<u32> {
-        let ctor_name = name?;
-        if let Ty::Named(type_name, _) = result_ty {
-            if let Some(cases) = self.emitter.variant_info.get(type_name.as_str()) {
-                for case in cases {
-                    if case.name == ctor_name {
-                        return Some(case.tag);
-                    }
-                }
-            }
-        }
-        None
-    }
-
     /// Emit a store instruction for a value at base_ptr + offset.
     /// Assumes base_ptr is already on stack, followed by the value.
-    fn emit_store_at(&mut self, ty: &Ty, offset: u32) {
+    pub fn emit_store_at(&mut self, ty: &Ty, offset: u32) {
         match values::ty_to_valtype(ty) {
             Some(ValType::I64) => {
                 self.func.instruction(&Instruction::I64Store(MemArg {
@@ -1309,181 +581,6 @@ impl FuncCompiler<'_> {
                 }));
             }
             _ => {}
-        }
-    }
-
-    /// Emit spread record: copy base, then overwrite specified fields.
-    fn emit_spread_record(&mut self, base: &IrExpr, overrides: &[(String, IrExpr)], result_ty: &Ty) {
-        let all_fields = self.extract_record_fields(result_ty);
-        let tag_offset = self.variant_tag_offset(result_ty);
-        let total_size = tag_offset + values::record_size(&all_fields);
-
-        // Allocate new record
-        self.func.instruction(&Instruction::I32Const(total_size as i32));
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-        let result_scratch = self.match_i32_base + self.match_depth;
-        self.func.instruction(&Instruction::LocalSet(result_scratch));
-
-        // Evaluate base and store ptr
-        self.emit_expr(base);
-        let base_scratch = result_scratch + 1;
-        self.func.instruction(&Instruction::LocalSet(base_scratch));
-
-        // Copy all bytes from base to result (including tag if variant)
-        // Byte-by-byte copy loop
-        // Use i64 scratch as counter
-        let counter = self.match_i64_base + self.match_depth;
-        self.func.instruction(&Instruction::I64Const(0));
-        self.func.instruction(&Instruction::LocalSet(counter));
-        self.func.instruction(&Instruction::Block(BlockType::Empty));
-        self.func.instruction(&Instruction::Loop(BlockType::Empty));
-        // break if counter >= total_size
-        self.func.instruction(&Instruction::LocalGet(counter));
-        self.func.instruction(&Instruction::I64Const(total_size as i64));
-        self.func.instruction(&Instruction::I64GeU);
-        self.func.instruction(&Instruction::BrIf(1));
-        // dst[i] = src[i]
-        self.func.instruction(&Instruction::LocalGet(result_scratch));
-        self.func.instruction(&Instruction::LocalGet(counter));
-        self.func.instruction(&Instruction::I32WrapI64);
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::LocalGet(base_scratch));
-        self.func.instruction(&Instruction::LocalGet(counter));
-        self.func.instruction(&Instruction::I32WrapI64);
-        self.func.instruction(&Instruction::I32Add);
-        self.func.instruction(&Instruction::I32Load8U(MemArg { offset: 0, align: 0, memory_index: 0 }));
-        self.func.instruction(&Instruction::I32Store8(MemArg { offset: 0, align: 0, memory_index: 0 }));
-        // counter++
-        self.func.instruction(&Instruction::LocalGet(counter));
-        self.func.instruction(&Instruction::I64Const(1));
-        self.func.instruction(&Instruction::I64Add);
-        self.func.instruction(&Instruction::LocalSet(counter));
-        self.func.instruction(&Instruction::Br(0));
-        self.func.instruction(&Instruction::End);
-        self.func.instruction(&Instruction::End);
-
-        // Overwrite specified fields
-        for (field_name, field_expr) in overrides {
-            if let Some((offset, _)) = values::field_offset(&all_fields, field_name) {
-                let total_offset = tag_offset + offset;
-                self.func.instruction(&Instruction::LocalGet(result_scratch));
-                self.emit_expr(field_expr);
-                self.emit_store_at(&field_expr.ty, total_offset);
-            }
-        }
-
-        // Return result ptr
-        self.func.instruction(&Instruction::LocalGet(result_scratch));
-    }
-
-    /// Emit a list literal: allocate [len:i32][elem0][elem1]...
-    fn emit_list(&mut self, elements: &[IrExpr], _list_ty: &Ty) {
-        let elem_ty = if let Some(first) = elements.first() {
-            first.ty.clone()
-        } else {
-            Ty::Int // empty list fallback
-        };
-        let elem_size = values::byte_size(&elem_ty);
-        let n = elements.len() as u32;
-        let total = 4 + n * elem_size;
-
-        self.func.instruction(&Instruction::I32Const(total as i32));
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-
-        let scratch = self.match_i32_base + self.match_depth;
-        self.func.instruction(&Instruction::LocalSet(scratch));
-
-        // Store length
-        self.func.instruction(&Instruction::LocalGet(scratch));
-        self.func.instruction(&Instruction::I32Const(n as i32));
-        self.func.instruction(&Instruction::I32Store(MemArg {
-            offset: 0, align: 2, memory_index: 0,
-        }));
-
-        // Store each element
-        for (i, elem) in elements.iter().enumerate() {
-            let offset = 4 + (i as u32) * elem_size;
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.emit_expr(elem);
-            self.emit_store_at(&elem.ty, offset);
-        }
-
-        self.func.instruction(&Instruction::LocalGet(scratch));
-    }
-
-    /// Emit index access: list_ptr + 4 + index * elem_size
-    fn emit_index_access(&mut self, object: &IrExpr, index: &IrExpr, result_ty: &Ty) {
-        let elem_size = values::byte_size(result_ty);
-
-        self.emit_expr(object); // list ptr
-        self.func.instruction(&Instruction::I32Const(4)); // skip len
-        self.func.instruction(&Instruction::I32Add);
-
-        // Add index * elem_size
-        self.emit_expr(index);
-        // Index might be i64 (Int), convert to i32
-        if matches!(&index.ty, Ty::Int) {
-            self.func.instruction(&Instruction::I32WrapI64);
-        }
-        self.func.instruction(&Instruction::I32Const(elem_size as i32));
-        self.func.instruction(&Instruction::I32Mul);
-        self.func.instruction(&Instruction::I32Add);
-
-        // Load element
-        self.emit_load_at(result_ty, 0);
-    }
-
-    /// Emit a tuple construction: allocate memory, store each element sequentially.
-    fn emit_tuple(&mut self, elements: &[IrExpr]) {
-        let element_types: Vec<(String, Ty)> = elements.iter().enumerate()
-            .map(|(i, e)| (format!("_{}", i), e.ty.clone()))
-            .collect();
-        let total_size = values::record_size(&element_types);
-
-        self.func.instruction(&Instruction::I32Const(total_size as i32));
-        self.func.instruction(&Instruction::Call(self.emitter.rt.alloc));
-
-        let scratch = self.match_i32_base + self.match_depth;
-        self.func.instruction(&Instruction::LocalSet(scratch));
-
-        let mut offset = 0u32;
-        for elem in elements {
-            let size = values::byte_size(&elem.ty);
-            self.func.instruction(&Instruction::LocalGet(scratch));
-            self.emit_expr(elem);
-            self.emit_store_at(&elem.ty, offset);
-            offset += size;
-        }
-
-        self.func.instruction(&Instruction::LocalGet(scratch));
-    }
-
-    /// Emit a tuple index access: load from tuple pointer + element offset.
-    fn emit_tuple_index(&mut self, object: &IrExpr, index: usize, result_ty: &Ty) {
-        // Compute offset by summing sizes of elements before `index`
-        let offset = if let Ty::Tuple(elem_types) = &object.ty {
-            elem_types.iter().take(index).map(|t| values::byte_size(t)).sum::<u32>()
-        } else {
-            0
-        };
-
-        self.emit_expr(object);
-        self.emit_load_at(result_ty, offset);
-    }
-
-    /// Emit a field access: load from record/variant pointer + field offset.
-    fn emit_member(&mut self, object: &IrExpr, field: &str) {
-        let fields = self.extract_record_fields(&object.ty);
-        // If the object is a variant type, fields start after the tag (offset +4)
-        let tag_offset = self.variant_tag_offset(&object.ty);
-
-        self.emit_expr(object); // ptr on stack
-
-        if let Some((field_offset, field_ty)) = values::field_offset(&fields, field) {
-            let total_offset = tag_offset + field_offset;
-            self.emit_load_at(&field_ty, total_offset);
-        } else {
-            self.func.instruction(&Instruction::Unreachable);
         }
     }
 
@@ -1510,7 +607,7 @@ impl FuncCompiler<'_> {
     }
 
     /// Returns 4 if the type is a variant (fields start after tag), 0 otherwise.
-    fn variant_tag_offset(&self, ty: &Ty) -> u32 {
+    pub(super) fn variant_tag_offset(&self, ty: &Ty) -> u32 {
         if let Ty::Named(name, _) = ty {
             if self.emitter.variant_info.contains_key(name.as_str()) {
                 return 4;
@@ -1524,7 +621,7 @@ impl FuncCompiler<'_> {
     }
 
     /// Extract field names and types from a record/named type.
-    fn extract_record_fields(&self, ty: &Ty) -> Vec<(String, Ty)> {
+    pub(super) fn extract_record_fields(&self, ty: &Ty) -> Vec<(String, Ty)> {
         match ty {
             Ty::Record { fields } => fields.clone(),
             Ty::Named(name, _) => {
@@ -1540,539 +637,8 @@ impl FuncCompiler<'_> {
         }
     }
 
-    /// Emit a for...in loop. Currently supports Range iterables only.
-    fn emit_for_in(&mut self, var: crate::ir::VarId, var_tuple: Option<&[crate::ir::VarId]>, iterable: &IrExpr, body: &[crate::ir::IrStmt]) {
-        match &iterable.kind {
-            IrExprKind::Range { start, end, inclusive } => {
-                let loop_var = self.var_map[&var.0];
-
-                // Initialize loop variable to start
-                self.emit_expr(start);
-                self.func.instruction(&Instruction::LocalSet(loop_var));
-
-                // block $break { loop $loop { check; block $continue { body }; i++; br $loop } }
-                let break_depth = self.depth;
-                self.func.instruction(&Instruction::Block(BlockType::Empty));
-                self.depth += 1;
-
-                let loop_depth = self.depth;
-                self.func.instruction(&Instruction::Loop(BlockType::Empty));
-                self.depth += 1;
-
-                // Break condition
-                self.func.instruction(&Instruction::LocalGet(loop_var));
-                self.emit_expr(end);
-                if *inclusive {
-                    self.func.instruction(&Instruction::I64GtS);
-                } else {
-                    self.func.instruction(&Instruction::I64GeS);
-                }
-                self.func.instruction(&Instruction::BrIf(self.depth - break_depth - 1));
-
-                // Inner block for continue target
-                let continue_depth = self.depth;
-                self.func.instruction(&Instruction::Block(BlockType::Empty));
-                self.depth += 1;
-
-                self.loop_stack.push(super::LoopLabels { break_depth, continue_depth });
-
-                // Body
-                for stmt in body {
-                    self.emit_stmt(stmt);
-                }
-
-                self.loop_stack.pop();
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end continue block
-
-                // Increment: var += 1 (always runs, even after continue)
-                self.func.instruction(&Instruction::LocalGet(loop_var));
-                self.func.instruction(&Instruction::I64Const(1));
-                self.func.instruction(&Instruction::I64Add);
-                self.func.instruction(&Instruction::LocalSet(loop_var));
-
-                // Loop back
-                self.func.instruction(&Instruction::Br(self.depth - loop_depth - 1));
-
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end loop
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end break block
-            }
-            _ => {
-                // List (or other collection) for...in
-                // scratch[0] = list ptr, scratch[1] = index counter
-                let list_scratch = self.match_i32_base + self.match_depth;
-                let idx_scratch = list_scratch + 1;
-                let loop_var = self.var_map[&var.0];
-
-                // Determine element type and size
-                let elem_ty = self.var_table.get(var).ty.clone();
-                let elem_size = values::byte_size(&elem_ty);
-
-                // Evaluate iterable and store list ptr
-                self.emit_expr(iterable);
-                self.func.instruction(&Instruction::LocalSet(list_scratch));
-
-                // Initialize index = 0
-                self.func.instruction(&Instruction::I32Const(0));
-                self.func.instruction(&Instruction::LocalSet(idx_scratch));
-
-                // Structure: block $break { loop $loop { check; load; block $continue { body }; i++; br $loop } }
-                // continue → br to $continue end (skips rest of body, runs i++)
-                let break_depth = self.depth;
-                self.func.instruction(&Instruction::Block(BlockType::Empty));
-                self.depth += 1;
-
-                let loop_depth = self.depth;
-                self.func.instruction(&Instruction::Loop(BlockType::Empty));
-                self.depth += 1;
-
-                // Break if index >= len
-                self.func.instruction(&Instruction::LocalGet(idx_scratch));
-                self.func.instruction(&Instruction::LocalGet(list_scratch));
-                self.func.instruction(&Instruction::I32Load(MemArg {
-                    offset: 0, align: 2, memory_index: 0,
-                }));
-                self.func.instruction(&Instruction::I32GeU);
-                self.func.instruction(&Instruction::BrIf(self.depth - break_depth - 1));
-
-                // Load element
-                self.func.instruction(&Instruction::LocalGet(list_scratch));
-                self.func.instruction(&Instruction::I32Const(4));
-                self.func.instruction(&Instruction::I32Add);
-                self.func.instruction(&Instruction::LocalGet(idx_scratch));
-                self.func.instruction(&Instruction::I32Const(elem_size as i32));
-                self.func.instruction(&Instruction::I32Mul);
-                self.func.instruction(&Instruction::I32Add);
-                self.emit_load_at(&elem_ty, 0);
-                self.func.instruction(&Instruction::LocalSet(loop_var));
-
-                // Tuple destructure
-                if let Some(tuple_vars) = var_tuple {
-                    if let Ty::Tuple(elem_types) = &elem_ty {
-                        let mut field_offset = 0u32;
-                        for (i, &tv) in tuple_vars.iter().enumerate() {
-                            if let Some(&local_idx) = self.var_map.get(&tv.0) {
-                                let ft = elem_types.get(i).cloned().unwrap_or(Ty::Int);
-                                self.func.instruction(&Instruction::LocalGet(loop_var));
-                                self.emit_load_at(&ft, field_offset);
-                                self.func.instruction(&Instruction::LocalSet(local_idx));
-                                field_offset += values::byte_size(&ft);
-                            }
-                        }
-                    }
-                }
-
-                // Inner block for continue target
-                let continue_depth = self.depth;
-                self.func.instruction(&Instruction::Block(BlockType::Empty));
-                self.depth += 1;
-
-                self.loop_stack.push(super::LoopLabels { break_depth, continue_depth });
-
-                // Body
-                for stmt in body {
-                    self.emit_stmt(stmt);
-                }
-
-                self.loop_stack.pop();
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end continue block
-
-                // Increment index (always runs, even after continue)
-                self.func.instruction(&Instruction::LocalGet(idx_scratch));
-                self.func.instruction(&Instruction::I32Const(1));
-                self.func.instruction(&Instruction::I32Add);
-                self.func.instruction(&Instruction::LocalSet(idx_scratch));
-
-                // Loop back
-                self.func.instruction(&Instruction::Br(self.depth - loop_depth - 1));
-
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end loop
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End); // end break block
-            }
-        }
-    }
-
-    /// Emit a match expression as a chain of if-else checks.
-    ///
-    /// Strategy: store subject in a scratch local, then for each arm:
-    /// - Literal pattern: compare subject to literal, branch if equal
-    /// - Wildcard: unconditional (last arm)
-    /// - Bind: store subject in the bound variable's local, unconditional
-    fn emit_match(&mut self, subject: &IrExpr, arms: &[IrMatchArm], result_ty: &Ty) {
-        // Resolve subject type — IR may have wrong type on Var nodes (type inference gap)
-        let subject_ty = self.resolve_subject_type(subject, arms);
-
-        let scratch = match values::ty_to_valtype(&subject_ty) {
-            Some(ValType::I64) => self.match_i64_base + self.match_depth,
-            _ => self.match_i32_base + self.match_depth,
-        };
-        self.match_depth += 1;
-
-        self.emit_expr(subject);
-        self.func.instruction(&Instruction::LocalSet(scratch));
-
-        self.emit_match_arms(arms, scratch, &subject_ty, result_ty, 0);
-
-        self.match_depth -= 1;
-    }
-
-    /// Resolve the actual subject type, fixing IR type inference gaps.
-    /// If the subject Var has the wrong type but patterns indicate a container type, fix it.
-    fn resolve_subject_type(&self, subject: &IrExpr, arms: &[IrMatchArm]) -> Ty {
-        let ty = &subject.ty;
-        // If patterns are Ok/Err/Some/None but subject type isn't a container, look up VarTable
-        let has_container_pattern = arms.iter().any(|a| matches!(
-            &a.pattern,
-            IrPattern::Ok { .. } | IrPattern::Err { .. } | IrPattern::Some { .. } | IrPattern::None
-        ));
-        if has_container_pattern && !matches!(ty, Ty::Applied(_, _)) {
-            // Try to get the real type from VarTable
-            if let IrExprKind::Var { id } = &subject.kind {
-                let info = self.var_table.get(*id);
-                if matches!(&info.ty, Ty::Applied(_, _)) {
-                    return info.ty.clone();
-                }
-            }
-        }
-        ty.clone()
-    }
-
-    fn emit_match_arms(
-        &mut self,
-        arms: &[IrMatchArm],
-        scratch: u32,
-        subject_ty: &Ty,
-        result_ty: &Ty,
-        idx: usize,
-    ) {
-        if idx >= arms.len() {
-            // No arms matched — should not happen with exhaustive match
-            self.func.instruction(&Instruction::Unreachable);
-            return;
-        }
-
-        let arm = &arms[idx];
-        let is_last = idx + 1 >= arms.len();
-
-        match &arm.pattern {
-            // Wildcard: always matches, emit body directly
-            IrPattern::Wildcard => {
-                self.emit_expr(&arm.body);
-            }
-
-            // Bind: store subject in variable, then emit body
-            IrPattern::Bind { var } => {
-                if let Some(&local_idx) = self.var_map.get(&var.0) {
-                    let var_ty = &self.var_table.get(*var).ty;
-                    let var_vt = values::ty_to_valtype(var_ty);
-                    let subj_vt = values::ty_to_valtype(subject_ty);
-                    // Only bind if types match, or var type is Unknown (trust subject)
-                    if var_vt == subj_vt || matches!(var_ty, Ty::Unknown) {
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.func.instruction(&Instruction::LocalSet(local_idx));
-                    }
-                }
-                self.emit_expr(&arm.body);
-            }
-
-            // Literal: compare subject to literal, if-else
-            IrPattern::Literal { expr: lit_expr } => {
-                // Push subject
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                // Push literal
-                self.emit_expr(lit_expr);
-                // Compare
-                match subject_ty {
-                    Ty::Int => { self.func.instruction(&Instruction::I64Eq); }
-                    Ty::Float => { self.func.instruction(&Instruction::F64Eq); }
-                    Ty::Bool => { self.func.instruction(&Instruction::I32Eq); }
-                    Ty::String => {
-                        // String equality: compare pointers (interned literals are deduped)
-                        self.func.instruction(&Instruction::I32Eq);
-                    }
-                    _ => { self.func.instruction(&Instruction::I32Eq); }
-                }
-
-                let bt = values::block_type(result_ty);
-                self.func.instruction(&Instruction::If(bt));
-                self.depth += 1;
-                self.emit_expr(&arm.body);
-                self.func.instruction(&Instruction::Else);
-
-                if is_last {
-                    self.func.instruction(&Instruction::Unreachable);
-                } else {
-                    self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1);
-                }
-
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End);
-            }
-
-            // Constructor pattern (e.g., Red, Some(x), None)
-            IrPattern::Constructor { name: ctor_name, args: _ } => {
-                // Look up tag for this constructor
-                if let Some(tag_val) = self.find_variant_tag_by_ctor(ctor_name, subject_ty) {
-                    // Load tag from subject and compare
-                    self.func.instruction(&Instruction::LocalGet(scratch));
-                    self.func.instruction(&Instruction::I32Load(MemArg {
-                        offset: 0, align: 2, memory_index: 0,
-                    }));
-                    self.func.instruction(&Instruction::I32Const(tag_val as i32));
-                    self.func.instruction(&Instruction::I32Eq);
-
-                    let bt = values::block_type(result_ty);
-                    self.func.instruction(&Instruction::If(bt));
-                    self.depth += 1;
-                    self.emit_expr(&arm.body);
-                    self.func.instruction(&Instruction::Else);
-                    if is_last {
-                        self.func.instruction(&Instruction::Unreachable);
-                    } else {
-                        self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1);
-                    }
-                    self.depth -= 1;
-                    self.func.instruction(&Instruction::End);
-                } else if is_last {
-                    self.emit_expr(&arm.body);
-                } else {
-                    self.func.instruction(&Instruction::Unreachable);
-                }
-            }
-
-            // Some(x) pattern (Option)
-            IrPattern::Some { inner } => {
-                // some(x) is a non-null pointer. Check ptr != 0, then load value.
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Const(0));
-                self.func.instruction(&Instruction::I32Ne);
-                let bt = values::block_type(result_ty);
-                self.func.instruction(&Instruction::If(bt));
-                self.depth += 1;
-
-                // Bind the inner value (type from subject, not VarTable — may be Unknown)
-                if let IrPattern::Bind { var } = inner.as_ref() {
-                    if let Some(&local_idx) = self.var_map.get(&var.0) {
-                        let inner_ty = if let Ty::Applied(_, args) = subject_ty {
-                            args.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.emit_load_at(&inner_ty, 0);
-                        self.func.instruction(&Instruction::LocalSet(local_idx));
-                    }
-                }
-
-                self.emit_expr(&arm.body);
-                self.func.instruction(&Instruction::Else);
-                if is_last {
-                    self.func.instruction(&Instruction::Unreachable);
-                } else {
-                    self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1);
-                }
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End);
-            }
-
-            // None pattern (Option)
-            IrPattern::None => {
-                // None is represented as i32 0
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Eqz);
-                let bt = values::block_type(result_ty);
-                self.func.instruction(&Instruction::If(bt));
-                self.depth += 1;
-                self.emit_expr(&arm.body);
-                self.func.instruction(&Instruction::Else);
-                if is_last {
-                    self.func.instruction(&Instruction::Unreachable);
-                } else {
-                    self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1);
-                }
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End);
-            }
-
-            // RecordPattern: variant constructor match (e.g., Circle { radius })
-            IrPattern::RecordPattern { name: ctor_name, fields: pat_fields, .. } => {
-                // Look up the tag for this constructor
-                let tag = self.find_variant_tag_by_ctor(ctor_name, subject_ty);
-
-                if let Some(tag_val) = tag {
-                    // Load tag from subject pointer
-                    self.func.instruction(&Instruction::LocalGet(scratch));
-                    self.func.instruction(&Instruction::I32Load(MemArg {
-                        offset: 0, align: 2, memory_index: 0,
-                    }));
-                    self.func.instruction(&Instruction::I32Const(tag_val as i32));
-                    self.func.instruction(&Instruction::I32Eq);
-
-                    let bt = values::block_type(result_ty);
-                    self.func.instruction(&Instruction::If(bt));
-                    self.depth += 1;
-
-                    // Bind fields: load each field from subject + tag_offset + field_offset
-                    let case_fields = self.emitter.record_fields.get(ctor_name).cloned().unwrap_or_default();
-                    for pf in pat_fields {
-                        // Find the field in the case's fields
-                        if let Some((foff, fty)) = values::field_offset(&case_fields, &pf.name) {
-                            let total_offset = 4 + foff; // 4 = tag size
-                            // Look up VarId for this field name in var_map
-                            // The pattern binds to a var with the same name
-                            // We need to find the VarId — it should be in var_map
-                            // The IR guarantees pattern fields create bindings in var_table
-                            // with the field name. We search by checking all var_map entries.
-                            // Actually, the var_table is indexed by VarId and has names.
-                            // We need to find the VarId that was allocated for this field name.
-                            // The scan_pattern in statements.rs should have registered it.
-                            // For now, find the local by searching var_map for the right VarId.
-
-                            // Simple approach: find the VarId from var_map whose name matches
-                            // This is set up by scan_pattern which registers field bindings
-                            if let Some(&local_idx) = self.find_var_by_field(&pf.name, &case_fields) {
-                                self.func.instruction(&Instruction::LocalGet(scratch));
-                                self.emit_load_at(&fty, total_offset);
-                                self.func.instruction(&Instruction::LocalSet(local_idx));
-                            }
-                        }
-                    }
-
-                    self.emit_expr(&arm.body);
-                    self.func.instruction(&Instruction::Else);
-
-                    if is_last {
-                        self.func.instruction(&Instruction::Unreachable);
-                    } else {
-                        self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1);
-                    }
-
-                    self.depth -= 1;
-                    self.func.instruction(&Instruction::End);
-                } else {
-                    // Not a variant — treat as plain record (always matches)
-                    self.emit_expr(&arm.body);
-                }
-            }
-
-            // Ok(x) pattern (Result)
-            IrPattern::Ok { inner } => {
-                // Result ok = tag 0. Check tag, then bind value.
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-                self.func.instruction(&Instruction::I32Eqz); // tag == 0
-                let bt = values::block_type(result_ty);
-                self.func.instruction(&Instruction::If(bt));
-                self.depth += 1;
-                if let IrPattern::Bind { var } = inner.as_ref() {
-                    if let Some(&local_idx) = self.var_map.get(&var.0) {
-                        // Ok inner type = args[0] of Result[T, E]
-                        let inner_ty = if let Ty::Applied(_, args) = subject_ty {
-                            args.first().cloned().unwrap_or(Ty::Int)
-                        } else { Ty::Int };
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.emit_load_at(&inner_ty, 4);
-                        self.func.instruction(&Instruction::LocalSet(local_idx));
-                    }
-                }
-                self.emit_expr(&arm.body);
-                self.func.instruction(&Instruction::Else);
-                if is_last { self.func.instruction(&Instruction::Unreachable); }
-                else { self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1); }
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End);
-            }
-
-            // Err(e) pattern (Result)
-            IrPattern::Err { inner } => {
-                self.func.instruction(&Instruction::LocalGet(scratch));
-                self.func.instruction(&Instruction::I32Load(MemArg { offset: 0, align: 2, memory_index: 0 }));
-                self.func.instruction(&Instruction::I32Const(0));
-                self.func.instruction(&Instruction::I32Ne);
-                let bt = values::block_type(result_ty);
-                self.func.instruction(&Instruction::If(bt));
-                self.depth += 1;
-                if let IrPattern::Bind { var } = inner.as_ref() {
-                    if let Some(&local_idx) = self.var_map.get(&var.0) {
-                        // Err inner type = args[1] of Result[T, E]
-                        let inner_ty = if let Ty::Applied(_, args) = subject_ty {
-                            args.get(1).cloned().unwrap_or(Ty::String)
-                        } else { Ty::String };
-                        self.func.instruction(&Instruction::LocalGet(scratch));
-                        self.emit_load_at(&inner_ty, 4);
-                        self.func.instruction(&Instruction::LocalSet(local_idx));
-                    }
-                }
-                self.emit_expr(&arm.body);
-                self.func.instruction(&Instruction::Else);
-                if is_last { self.func.instruction(&Instruction::Unreachable); }
-                else { self.emit_match_arms(arms, scratch, subject_ty, result_ty, idx + 1); }
-                self.depth -= 1;
-                self.func.instruction(&Instruction::End);
-            }
-
-            // Tuple pattern: (a, b) => ...
-            IrPattern::Tuple { elements } => {
-                // Tuple always matches (destructure only). Bind each element.
-                if let Ty::Tuple(elem_types) = subject_ty {
-                    let mut offset = 0u32;
-                    for (i, elem_pat) in elements.iter().enumerate() {
-                        if let IrPattern::Bind { var } = elem_pat {
-                            if let Some(&local_idx) = self.var_map.get(&var.0) {
-                                let ft = elem_types.get(i).cloned().unwrap_or(Ty::Int);
-                                self.func.instruction(&Instruction::LocalGet(scratch));
-                                self.emit_load_at(&ft, offset);
-                                self.func.instruction(&Instruction::LocalSet(local_idx));
-                                offset += values::byte_size(&ft);
-                            }
-                        } else if let IrPattern::Wildcard = elem_pat {
-                            let ft = elem_types.get(i).cloned().unwrap_or(Ty::Int);
-                            offset += values::byte_size(&ft);
-                        }
-                    }
-                }
-                self.emit_expr(&arm.body);
-            }
-
-            // Catch-all for unsupported patterns
-            _ => {
-                if is_last {
-                    self.emit_expr(&arm.body);
-                } else {
-                    self.func.instruction(&Instruction::Unreachable);
-                }
-            }
-        }
-    }
-
-    /// Find the variant tag for a constructor name, searching variant_info by subject type.
-    fn find_variant_tag_by_ctor(&self, ctor_name: &str, subject_ty: &Ty) -> Option<u32> {
-        let type_name = match subject_ty {
-            Ty::Named(name, _) => name.as_str(),
-            Ty::Variant { name, .. } => name.as_str(),
-            _ => return None,
-        };
-        let cases = self.emitter.variant_info.get(type_name)?;
-        cases.iter().find(|c| c.name == ctor_name).map(|c| c.tag)
-    }
-
-    /// Find variant tag for a unit constructor called as a function (e.g., `Red`).
-    fn find_unit_variant_tag(&self, name: &str) -> Option<u32> {
-        for cases in self.emitter.variant_info.values() {
-            for case in cases {
-                if case.name == name && case.fields.is_empty() {
-                    return Some(case.tag);
-                }
-            }
-        }
-        None
-    }
-
     /// Find local index for a pattern field binding by name.
-    fn find_var_by_field(&self, field_name: &str, _case_fields: &[(String, Ty)]) -> Option<&u32> {
+    pub(super) fn find_var_by_field(&self, field_name: &str, _case_fields: &[(String, Ty)]) -> Option<&u32> {
         // Search var_map for VarIds whose name in var_table matches field_name
         for (&var_id, local_idx) in &self.var_map {
             if (var_id as usize) < self.var_table.len() {
@@ -2084,12 +650,27 @@ impl FuncCompiler<'_> {
         }
         None
     }
-}
 
-#[derive(Clone, Copy)]
-enum CmpKind {
-    Lt,
-    Gt,
-    Lte,
-    Gte,
+    /// Find variant tag for a unit constructor called as a function (e.g., `Red`).
+    pub(super) fn find_unit_variant_tag(&self, name: &str) -> Option<u32> {
+        for cases in self.emitter.variant_info.values() {
+            for case in cases {
+                if case.name == name && case.fields.is_empty() {
+                    return Some(case.tag);
+                }
+            }
+        }
+        None
+    }
+
+    /// Find the variant tag for a constructor name, searching variant_info by subject type.
+    pub(super) fn find_variant_tag_by_ctor(&self, ctor_name: &str, subject_ty: &Ty) -> Option<u32> {
+        let type_name = match subject_ty {
+            Ty::Named(name, _) => name.as_str(),
+            Ty::Variant { name, .. } => name.as_str(),
+            _ => return None,
+        };
+        let cases = self.emitter.variant_info.get(type_name)?;
+        cases.iter().find(|c| c.name == ctor_name).map(|c| c.tag)
+    }
 }
