+++ src/codegen/emit_wasm/mod.rs
@@ -16,6 +16,9 @@ pub mod values;
 mod strings;
 mod runtime;
 mod expressions;
+mod calls;
+mod collections;
+mod control;
 pub mod statements;
 mod functions;
